requirements.txt
absolufy-imports==0.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.10.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.5
auto-labeling-pipeline==0.1.23
babel==2.17.0
backports-datetime-fromisoformat==2.0.3
beautifulsoup4==4.14.2
billiard==4.2.2
blake3==1.0.7
bleach==6.2.0
blis==1.3.0
boto3==1.40.55
botocore==1.40.55
CacheControl==0.14.3
cachetools==5.5.2
catalogue==2.0.10
cbor2==5.7.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.21.1
cloudpickle==3.1.1
comm==0.2.3
compressed-tensors==0.11.0
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
datasets==4.1.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dj-database-url==0.5.0
dj-rest-auth==2.2.8
Django==4.2.25
django-allauth==0.52.0
django-celery-results==2.4.0
django-cleanup==6.0.0
django-cors-headers==3.14.0
django-drf-filepond==0.5.0
django-filter==22.1
django-health-check==3.20.0
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework-xml==2.0.0
dnspython==2.7.0
docstring_parser==0.17.0
drf-yasg==1.21.11
dss-python==0.15.7
dss_python_backend==0.14.5
einops==0.8.1
email_validator==2.2.0
environs==9.5.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
fastjsonschema==2.21.2
fastuuid==0.13.5
filelock==3.19.1
filetype==1.2.0
firebase-admin==6.9.0
-e git+https://github.com/facebookresearch/flow_matching@a1ef82096594a5569fdc82561194386e41031f29#egg=flow_matching&subdirectory=../../flow_matching
flower==1.2.0
fonttools==4.60.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
furl==2.1.4
gguf==0.17.1
google==3.0.0
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.121.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigtable==2.33.0
google-cloud-core==2.4.3
google-cloud-discoveryengine==0.13.12
google-cloud-firestore==2.21.0
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.58.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.45.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==20.1.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
humanize==4.14.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
interegular==0.3.3
ipykernel==7.0.1
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
litellm==1.78.5
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
lml==0.2.0
lxml==6.0.2
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
matplotlib==3.7.1
matplotlib-inline==0.1.7
mcp==1.18.0
mdurl==0.1.2
mistral_common==1.8.5
mistune==3.1.4
mlx==0.29.2
mlx-lm==0.28.1
mlx-metal==0.29.2
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
murmurhash==1.0.13
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
openai==1.109.1
openai-harmony==0.0.4
opencv-python-headless==4.12.0.88
opendssdirect.py==0.9.4
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.10.0a0
opentelemetry-exporter-gcp-monitoring==1.10.0a0
opentelemetry-exporter-gcp-trace==1.10.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.10.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orderedmultidict==1.0.1
outlines_core==0.2.11
overrides==7.7.0
packaging==25.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.5
partial-json-parser==0.2.1.1.post6
pdfminer.six==20250506
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
preshed==3.0.10
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyexcel==0.7.3
pyexcel-io==0.6.7
pyexcel-xlsx==0.6.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.0.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.12.2
Send2Trash==1.8.3
sentencepiece==0.2.1
sentry-sdk==2.39.0
seqeval==1.2.2
setproctitle==1.3.7
shapely==2.1.2
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
sqlalchemy-spanner==1.17.0
sqlparse==0.5.3
srsly==2.5.1
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tenacity==8.5.0
terminado==0.18.1
texttable==1.7.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.8.0
torchaudio==2.8.0
torchdiffeq==0.2.5
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.21.0
vine==5.1.0
vllm==0.10.2
waitress==2.1.2
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.14
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
whitenoise==6.11.0
wrapt==1.17.2
xgrammar==0.1.23
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
